Upsell Cluster - New CSS and progress on new experience
Making progress. Checking in new CSS to fix some of the positioning
problems with the data in the new row. There is still a problem that
when you click on one of the radio buttons for the list of options that
the options jump to a margin-top: -177px and all room type rows open.