Triptease Module - rearranged order of events
I rearranged the order of events in this.  The HTML changes were being
applied to all rows, and then it tried to take out the ones that didn’t
need it.  I reversed that, so to speak.  I changed the order of the code
to not make any changes unless the hotel was in the array.  If it is,
then it changes just that row. I also put a counter in the each loop
that increments by one each time it finds an approved hotel. When the
counter reaches 6, it finishes the changes for that hotel and then exits
the each loop.