Special Offer Group - CSS update
Code works and moves the right rates to the right place.  Now adding
HTML and CSS styling to match the mockups. This check in has the top and
bottom buttons and they properly open and close the hidden area. Bottom
button needs to be moved to right spot, border color needs to be fixed
for hidden area, and SVG icons need to be added to buttons.