4 Rate Pack - added taoRatePack array
Added an array at the top of the script that allows you to designate
what rates are featured in the rate pack.  All code now updated to run
off of this list.  Featured rates are displayed in order they are added
to the array.  Array has functionality to add a new title for the rate
but this part hasn't been tested.